requirements.txt
Flask
flask_login
flask_sqlalchemy
flask_wtf
werkzeug
pymysql
pandas
openpyxl